cast da_ht to int when passed to variadic g_object_set
don't crash when opening a SF instance

+++ src/ui/scoring-facility.cc
@@ -199,7 +199,7 @@ SScoringFacility( agh::CSubject& J,
 		      "height-request", HypnogramHeight,
 		      NULL);
 	g_object_set( (GObject*)daSFMontage,
-		      "height-request", da_ht,
+		      "height-request", (int)da_ht,
 		      NULL);
